==> 03_features_collections.ipynb <==
import datetime
import geopandas as gpd
import matplotlib.pyplot as plt
import pandas as pd
import numpy as np
from shapely.geometry import Point
from geopandas import GeoDataFrame
import os

import seaborn as sns
pd.set_option("display.max_colwidth", 0)
pd.set_option("display.max_rows", 500)
pd.set_option("display.max_columns", 500)
pd.set_option("display.width", 1000)
pd.options.display.float_format = '{:.5f}'.format

import warnings
warnings.filterwarnings("ignore")

%%time
df_localness = pd.read_csv('./data/food_localness_oct.csv',header=0) #,nrows=10
# df_localness

df_localness['customer_loc_id'] = df_localness['customer_lon'].astype(str)+df_localness['customer_lat'].astype(str)

## Create weekend binary column

%%time
# Create one column to determine if it is weekend
df_localness['order_date'] = pd.to_datetime(df_localness['order_date']) 

# Define a function to check if a date falls on a weekend
def is_weekend(date):
    return (date.dayofweek == 5) or (date.dayofweek == 6)

# Apply the function to the 'Date' column
df_localness['isweekend'] = df_localness['order_date'].apply(lambda x: is_weekend(x))

## Merge with mean price per item (proxy for expensiveness of restaurants)

df_priceper_item = pd.read_csv('./data/mean_price_per_item.csv')
df_localness = df_localness.merge(df_priceper_item, how='left',on='vendor_id')
print(df_localness.shape[0])

df_localness = df_localness[df_localness['item_price']!=0]
print(df_localness.shape[0])

## merge entropy and restaurant count data

# %%time
# distance_list = [0.5, 1, 2, 3, 4, 5] #unit:km

# # left join various entropy indices and number of resturants
# for distance_threshold in distance_list:
#     entropy_col_name = 'entropy_'+str(distance_threshold)+'km'
#     poi_col_name = 'poi_count_'+str(distance_threshold)+'km'
#     entropy_file_path = './data/entropy_'+str(distance_threshold)+'km_customerlocations_oct.csv'
#     poicount_file_path = './data/poi_count_'+str(distance_threshold)+'km_customerlocations_oct.csv'
    
#     entropy_bycustomerloc1 = pd.read_csv(entropy_file_path,header=0)
#     df_localness = df_localness.merge(entropy_bycustomerloc1, how='left',on='customer_loc_id')
#     df_localness[entropy_col_name] = df_localness[entropy_col_name].fillna(0)
    
#     poi_bycustomerloc1 = pd.read_csv(poicount_file_path,header=0)
#     df_localness = df_localness.merge(poi_bycustomerloc1, how='left',on='customer_loc_id')
#     df_localness[poi_col_name] = df_localness[poi_col_name].fillna(0)
#     print(df_localness.shape[0])

%%time
distance_list = [0.5, 1, 2, 3, 4, 5] #unit:km

# left join various entropy indices and number of resturants
for distance_threshold in distance_list:
    entropy_col_name = 'entropy_'+str(distance_threshold)+'km'
    poi_col_name = 'poi_count_'+str(distance_threshold)+'km'
    entropy_file_path = './data/entropy_'+str(distance_threshold)+'km_customerlocations_oct.csv'
    poicount_file_path = './data/poi_count_'+str(distance_threshold)+'km_customerlocations_oct.csv'
    
    entropy_bycustomerloc1 = pd.read_csv(entropy_file_path,header=0)
    df_localness = df_localness.merge(entropy_bycustomerloc1, how='left',on='order_id')
    df_localness[entropy_col_name] = df_localness[entropy_col_name].fillna(0)
    
    poi_bycustomerloc1 = pd.read_csv(poicount_file_path,header=0)
    df_localness = df_localness.merge(poi_bycustomerloc1, how='left',on='order_id')
    df_localness[poi_col_name] = df_localness[poi_col_name].fillna(0)
    print(df_localness.shape[0])

df_localness['poi_count_0.5_1km'] = df_localness['poi_count_1km'] - df_localness['poi_count_0.5km']
df_localness['poi_count_1_2km'] = df_localness['poi_count_2km'] - df_localness['poi_count_1km']
df_localness['poi_count_2_3km'] = df_localness['poi_count_3km'] - df_localness['poi_count_2km']
df_localness['poi_count_3_4km'] = df_localness['poi_count_4km'] - df_localness['poi_count_3km']
df_localness['poi_count_4_5km'] = df_localness['poi_count_5km'] - df_localness['poi_count_4km']

# df_localness.head(3)

## Merge with rent data

unique_customerlocs_withrent = pd.read_csv('./data/estimated_rent_customerlocations_oct.csv')
print(unique_customerlocs_withrent.shape[0])
unique_customerlocs_withrent.head()

df_localness = df_localness.merge(unique_customerlocs_withrent, how='left',on='customer_loc_id')
print(df_localness.shape[0])

df_localness.head(2)

## merge incentives data

df_incentives = pd.read_csv('./data/mit_base_orders_incentives_oct.csv')
df_incentives.head(2)

%%time
df_localness = df_localness.merge(df_incentives, on='order_id',how='left')
print(df_localness.shape[0])
df_localness.fillna(0,inplace=True)

# create one categorical var for incentive types
df_localness['incentives_type'] = 'none'
df_localness.loc[df_localness['compensation'] != 0,'incentives_type'] = 'compensation'
df_localness.loc[df_localness['discounted delivery'] != 0,'incentives_type'] = 'discounted delivery'
df_localness.loc[df_localness['free delivery'] != 0,'incentives_type'] = 'free delivery'
df_localness.loc[df_localness['gem'] != 0,'incentives_type'] = 'gem'
df_localness.loc[df_localness['item discount'] != 0,'incentives_type'] = 'item discount'
df_localness.loc[df_localness['menu discount'] != 0,'incentives_type'] = 'menu discount'
df_localness.loc[df_localness['voucher wallet'] != 0,'incentives_type'] = 'voucher wallet'

# create three dummy vars for three most important incentives
df_localness['incentive_delivery'] = 'False'
df_localness.loc[df_localness['discounted delivery'] != 0,'incentive_delivery'] = 'True'
df_localness.loc[df_localness['free delivery'] != 0,'incentive_delivery'] = 'True'

df_localness['incentive_menu_discount'] = 'False'
df_localness.loc[df_localness['menu discount'] != 0,'incentive_menu_discount'] = 'True'

df_localness['incentive_voucher_wallet'] = 'False'
df_localness.loc[df_localness['voucher wallet'] != 0,'incentive_voucher_wallet'] = 'True'

## Merge with rating data

df_ratings = pd.read_csv('./data/restaurant_ratings.csv')
df_ratings = df_ratings[['vendor_id','rating']]
print(df_ratings.shape[0])
print(df_ratings.rating.value_counts())
df_ratings.head()

# left join with the picked resturants
df_localness = df_localness.merge(df_ratings, how='left')
print(df_localness.shape[0])
df_localness['rating'] = df_localness['rating'].fillna('None')
print(df_localness.shape[0])

# left join with nearest restaurants
df_ratings = df_ratings.rename(columns={"vendor_id": "n_vendor_id", "rating":"n_rating"})
df_localness = df_localness.merge(df_ratings, how='left')
df_localness['n_rating'] = df_localness['n_rating'].fillna('None')
print(df_localness.shape[0])

## Merge with individual order consistency indicator: entropy
Lower value of entropy means high consistency

df_consistency = pd.read_csv('./data/order_consistency_oct.csv')
print(df_consistency.shape[0])
df_consistency.head(2)

df_localness = df_localness.merge(df_consistency, how='left')
print(df_localness['consistency'].isnull().values.any())
print(df_localness.shape[0])

## Merge with total number of orders in one month for each individual
entropy of consistency itself may be biased so also incorporate total number of orders each person placed in Oct

df_totalorders = df_localness[['account_id','vendor_id']].groupby('account_id').count().reset_index()
df_totalorders.columns = ['account_id', 'total_orders']
# df_totalorders.head()
df_localness = df_localness.merge(df_totalorders, how='left')
print(df_localness['total_orders'].isnull().values.any())
print(df_localness.shape[0])

## Rename columns to avoid confusions

# exclude rows without rent data
print('no rent data points:',df_localness[df_localness.approxi_rent==0].shape[0])
df_localness_filter = df_localness[df_localness.approxi_rent!=0]
print(df_localness_filter.shape[0])

df_localness_filter.loc[df_localness_filter.rating == 'Amazing', 'rating'] = '1.Amazing'
df_localness_filter.loc[df_localness_filter.rating == 'Very good', 'rating'] = '2.Very good'
df_localness_filter.loc[df_localness_filter.rating == 'Good', 'rating'] = '3.Good'
df_localness_filter.loc[df_localness_filter.rating == 'OK', 'rating'] = '4.OK'
df_localness_filter.loc[df_localness_filter.rating == 'None', 'rating'] = '5.None'

df_localness_filter.loc[df_localness_filter.n_rating == 'Amazing', 'n_rating'] = '1.Amazing'
df_localness_filter.loc[df_localness_filter.n_rating == 'Very good', 'n_rating'] = '2.Very good'
df_localness_filter.loc[df_localness_filter.n_rating == 'Good', 'n_rating'] = '3.Good'
df_localness_filter.loc[df_localness_filter.n_rating == 'OK', 'n_rating'] = '4.OK'
df_localness_filter.loc[df_localness_filter.n_rating == 'None', 'n_rating'] = '5.None'

df_localness_filter.loc[df_localness_filter.isweekend == True, 'isweekend'] = 'Yes'
df_localness_filter.loc[df_localness_filter.isweekend == False, 'isweekend'] = 'No'

df_localness_filter.loc[df_localness_filter.incentive_delivery == True, 'incentive_delivery'] = 'Yes'
df_localness_filter.loc[df_localness_filter.incentive_delivery == False, 'incentive_delivery'] = 'No'

df_localness_filter.loc[df_localness_filter.incentive_menu_discount == True, 'incentive_menu_discount'] = 'Yes'
df_localness_filter.loc[df_localness_filter.incentive_menu_discount == False, 'incentive_menu_discount'] = 'No'

df_localness_filter.loc[df_localness_filter.incentive_voucher_wallet == True, 'incentive_voucher_wallet'] = 'Yes'
df_localness_filter.loc[df_localness_filter.incentive_voucher_wallet == False, 'incentive_voucher_wallet'] = 'No'

df_localness_filter = df_localness_filter.rename(columns={"item_price": "price_per_item", "approxi_rent":"estimated_rent",
  "poi_count_0.5km" : "restaurant_count_0.5km", "poi_count_0.5_1km": "restaurant_count_0.5_1km",
  "poi_count_1_2km": "restaurant_count_1_2km", "poi_count_2_3km" : "restaurant_count_2_3km", 
  "poi_count_3_4km" : "restaurant_count_3_4km", "poi_count_4_5km" : "restaurant_count_4_5km"
})

## Correct localness value

df_localness_filter.head(2)

df_localness_filter['nearest_dist1'] = df_localness_filter['nearest_dist1']+0.001 # plus one meter to correct zero localness when both pandana_distance_realdist and nearest_dist1 are zero
df_localness_filter['pandana_distance_realdist'] = df_localness_filter['pandana_distance_realdist']+0.001
df_localness_filter['pandana_localness'] = df_localness_filter['nearest_dist1']/df_localness_filter['pandana_distance_realdist']

print(df_localness_filter.shape[0])

## Save processed data for ML model

%%time
df_localness_filter.to_csv('./data/food_localness_final_oct_new.csv',index=False)

# %%time
# df_localness.to_csv('./data/food_localness_final_oct(withoutfilter_rent).csv',index=False)
